client-webservice/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.29.0

pydantic==1.10.14
python-dotenv==1.0.1

pymongo==4.6.0
httpx==0.27.0

email-validator
grpcio
grpcio-tools
